+++ packaging/nix/package.nix
@@ -21,7 +21,7 @@
 
 python3Packages.buildPythonApplication rec {
   pname = "hefesto-dualsense4unix";
-  version = "3.13.2";
+  version = "3.13.3";
   pyproject = true;
 
   # Source local (clonado pelo flake). Em release tag, trocar por